requirements.txt
pylint
scrapy
parsel
requests
parmap
tqdm
Flask
Flask-SQLAlchemy
pydocstyle
ijson
fuzzywuzzy[speedup]
cchardet
lxml
https://github.com/rogerbinns/apsw/releases/download/3.27.2-r1/apsw-3.27.2-r1.zip --global-option=fetch --global-option=--version --global-option=3.27.2 --global-option=--all --global-option=build --global-option=--enable-all-extensions
